ofb-flower/server/flower/requirements.txt
flwr==0.19.0
mlflow
torchvision
pathlib
torchsummary
numpy 
scikit-image
sklearn
imgaug
pytorch_lightning
pytorch-lightning-bolts
lightning-transformers
pysftp
timm